# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Werd API
  version: 0.0.1
  description: Werd platform API — multi-project social media automation and monitoring.
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8090
    description: Local development
paths:
...
  /projects/{id}/connections:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: List platform connections
      operationId: listConnections
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
    post:
      summary: Create platform connection
      operationId: createConnection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform]
              properties:
                platform:
                  type: string
                method:
                  type: string
                  enum: [api, browser]
                  default: api
                credentials:
                  type: object
      responses:
        "201":
          description: Connection created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
  /projects/{id}/connections/{connID}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: connID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get connection by ID
      operationId: getConnection
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connection details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
    put:
      summary: Update connection
      operationId: updateConnection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                method:
                  type: string
                  enum: [api, browser]
                credentials:
                  type: object
                enabled:
                  type: boolean
      responses:
        "200":
          description: Connection updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
    delete:
      summary: Delete connection
      operationId: deleteConnection
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connection deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /projects/{id}/posts:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: List posts
      operationId: listPosts
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, scheduled, publishing, published, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated post list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
    post:
      summary: Create post
      operationId: createPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, platforms]
              properties:
                title:
                  type: string
                content:
                  type: string
                url:
                  type: string
                post_type:
                  type: string
                  enum: [text, link]
                platforms:
                  type: array
                  items:
                    type: string
                reply_to_url:
                  type: string
                  description: URL of existing post/thread to reply to (single platform only)
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /projects/{id}/posts/{postID}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: postID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get post by ID
      operationId: getPost
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Post details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    put:
      summary: Update post
      operationId: updatePost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, platforms]
              properties:
                title:
                  type: string
                content:
                  type: string
                url:
                  type: string
                post_type:
                  type: string
                  enum: [text, link]
                platforms:
                  type: array
                  items:
                    type: string
                reply_to_url:
                  type: string
                  description: URL of existing post/thread to reply to (single platform only)
      responses:
        "200":
          description: Post updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
    delete:
      summary: Delete post
      operationId: deletePost
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Post deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /projects/{id}/posts/{postID}/publish:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: postID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Publish post to all target platforms
      operationId: publishPost
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Publish results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
  /projects/{id}/posts/{postID}/monitor:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: postID
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      summary: Enable/disable reply monitoring for a post
      operationId: setPostMonitor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enable]
              properties:
                enable:
                  type: boolean
      responses:
        "200":
          description: Monitor setting updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-Key
  schemas:
...
    Message:
      type: object
      properties:
        message:
          type: string
...
    Connection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        platform:
          type: string
        method:
          type: string
          enum: [api, browser]
        target:
          type: string
          description: Non-secret display target (e.g. r/golang, @user.bsky.social)
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        url:
          type: string
        post_type:
          type: string
          enum: [text, link]
        platforms:
          type: array
          items:
            type: string
        reply_to_url:
          type: string
          description: URL of existing post/thread this is a reply to
        scheduled_at:
          type: string
          format: date-time
          nullable: true
        published_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [draft, scheduled, publishing, published, failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PostListResponse:
      type: object
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        total:
          type: integer
    PlatformPublishResult:
      type: object
      properties:
        platform:
          type: string
        success:
          type: boolean
        post_id:
          type: string
        url:
          type: string
        error:
          type: string
    PublishResponse:
      type: object
      properties:
        post:
          $ref: "#/components/schemas/Post"
        results:
          type: array
          items:
            $ref: "#/components/schemas/PlatformPublishResult"
...